requirements.txt
# Pandas Dependencies
numpy>=1.7.1
# Pandas Optional Dependencies
scipy>=0.12.0
# Pandas
pandas>=0.12.0
# Tests
unittest2py3k
mock
